test(core): configure shared vitest setup
