+++ packages/core/test/actions/csrfToken/stateless.test.ts
@@ -1,13 +1,9 @@
-import { describe, test, expect, vi, beforeEach, afterEach } from "vitest"
+import { describe, test, expect, vi, beforeEach } from "vitest"
 import { GET } from "@test/presets.ts"
 import { setCookie } from "@/cookie.ts"
 
 beforeEach(() => {
     vi.stubEnv("BASE_URL", undefined)
-})
-
-afterEach(() => {
-    vi.unstubAllEnvs()
 })
 
 describe("csrfTokenAction", () => {
