+++ packages/core/test/actions/providers/tokens/tokens/stateless.test.ts
@@ -1,4 +1,4 @@
-import { describe, test, expect, vi, afterEach, beforeEach } from "vitest"
+import { describe, test, expect, vi, beforeEach } from "vitest"
 import { createAuth } from "@/createAuth.ts"
 import { createCSRF } from "@/shared/crypto.ts"
 import { GET, jose, oauthCustomService, oauthTokens, sessionPayload } from "@test/presets.ts"
@@ -7,12 +7,6 @@ import type { OAuthProviderConfig } from "@/@types/oauth.ts"
 
 beforeEach(() => {
     vi.stubEnv("BASE_URL", undefined)
-})
-
-afterEach(() => {
-    vi.unstubAllEnvs()
-    vi.restoreAllMocks()
-    vi.unstubAllGlobals()
 })
 
 describe("tokensAction", async () => {
