+++ packages/core/test/actions/signIn/signInCredentials/stateful.test.ts
@@ -1,39 +1,10 @@
-import { describe, test, expect, beforeEach, vi, afterEach } from "vitest"
+import { describe, test, expect, beforeEach, vi } from "vitest"
 import { authInstance, jose, sessionEntityWithUser, userEntity } from "@test/presets.ts"
 import { createCSRF } from "@/shared/crypto.ts"
 import { createSchemaRegistry } from "@/validator/registry.ts"
 
 beforeEach(() => {
     vi.stubEnv("BASE_URL", undefined)
-})
-
-afterEach(() => {
-    vi.unstubAllEnvs()
-    vi.restoreAllMocks()
-    vi.unstubAllGlobals()
-})
-
-vi.mock("@aura-stack/rate-limiter", async () => {
-    const actual = await vi.importActual<typeof import("@aura-stack/rate-limiter")>("@aura-stack/rate-limiter")
-    return {
-        ...actual,
-        createRateLimiter: (...args: Parameters<typeof actual.createRateLimiter>) => {
-            const limiters = actual.createRateLimiter(...args)
-
-            for (const limiter of Object.values(limiters)) {
-                limiter.check = vi.fn().mockResolvedValue({
-                    ok: true,
-                    limit: Number.MAX_SAFE_INTEGER,
-                    remaining: Number.MAX_SAFE_INTEGER,
-                    resetAt: Date.now() + 60000,
-                    retryAfter: 0,
-                    toResponse: () => new Response(),
-                })
-            }
-
-            return limiters
-        },
-    }
 })
 
 describe("signInCredentials action", async () => {
