+++ packages/core/test/actions/signUp/stateless.test.ts
@@ -1,4 +1,4 @@
-import { describe, test, expect, beforeEach, vi, afterEach } from "vitest"
+import { describe, test, expect, beforeEach, vi } from "vitest"
 import { z } from "zod/v4"
 import { jose, POST } from "@test/presets.ts"
 import { createAuth } from "@/createAuth.ts"
@@ -10,41 +10,12 @@ beforeEach(() => {
     vi.stubEnv("BASE_URL", undefined)
 })
 
-afterEach(() => {
-    vi.unstubAllEnvs()
-    vi.restoreAllMocks()
-    vi.unstubAllGlobals()
-})
-
 const payload = {
     name: "johndoe",
     email: "john@example.com",
     image: "https://example.com/image.jpg",
     password: "1234567890",
 }
-
-vi.mock("@aura-stack/rate-limiter", async () => {
-    const actual = await vi.importActual<typeof import("@aura-stack/rate-limiter")>("@aura-stack/rate-limiter")
-    return {
-        ...actual,
-        createRateLimiter: (...args: Parameters<typeof actual.createRateLimiter>) => {
-            const limiters = actual.createRateLimiter(...args)
-
-            for (const limiter of Object.values(limiters)) {
-                limiter.check = vi.fn().mockResolvedValue({
-                    ok: true,
-                    limit: Number.MAX_SAFE_INTEGER,
-                    remaining: Number.MAX_SAFE_INTEGER,
-                    resetAt: Date.now() + 60000,
-                    retryAfter: 0,
-                    toResponse: () => new Response(),
-                })
-            }
-
-            return limiters
-        },
-    }
-})
 
 describe("signUp API", async () => {
     const csrfToken = await createCSRF(jose)
