+++ packages/core/test/api/stateful/signOut.test.ts
@@ -1,38 +1,9 @@
-import { afterEach, beforeEach, describe, expect, test, vi } from "vitest"
+import { beforeEach, describe, expect, test, vi } from "vitest"
 import { createCSRF } from "@/shared/crypto.ts"
 import { authInstance, jose, sessionEntityWithUser } from "@test/presets.ts"
 
 beforeEach(() => {
     vi.stubEnv("BASE_URL", undefined)
-})
-
-afterEach(() => {
-    vi.unstubAllEnvs()
-    vi.restoreAllMocks()
-    vi.unstubAllGlobals()
-})
-
-vi.mock("@aura-stack/rate-limiter", async () => {
-    const actual = await vi.importActual<typeof import("@aura-stack/rate-limiter")>("@aura-stack/rate-limiter")
-    return {
-        ...actual,
-        createRateLimiter: (...args: Parameters<typeof actual.createRateLimiter>) => {
-            const limiters = actual.createRateLimiter(...args)
-
-            for (const limiter of Object.values(limiters)) {
-                limiter.check = vi.fn().mockResolvedValue({
-                    ok: true,
-                    limit: Number.MAX_SAFE_INTEGER,
-                    remaining: Number.MAX_SAFE_INTEGER,
-                    resetAt: Date.now() + 60000,
-                    retryAfter: 0,
-                    toResponse: () => new Response(),
-                })
-            }
-
-            return limiters
-        },
-    }
 })
 
 describe("signOut API", async () => {
