+++ packages/core/test/api/stateless/getAccessToken.test.ts
@@ -1,4 +1,4 @@
-import { describe, test, expect, vi, beforeEach, afterEach } from "vitest"
+import { describe, test, expect, vi, beforeEach } from "vitest"
 import { api, jose, oauthCustomService, oauthTokens, sessionPayload } from "@test/presets.ts"
 import { createCSRF } from "@/shared/crypto.ts"
 import { createAuth } from "@/createAuth.ts"
@@ -9,18 +9,12 @@ beforeEach(() => {
     vi.stubEnv("BASE_URL", undefined)
 })
 
-afterEach(() => {
-    vi.unstubAllEnvs()
-    vi.restoreAllMocks()
-    vi.unstubAllGlobals()
-})
-
-describe("getProviderTokens API", () => {
+describe("getAccessToken API", () => {
     test("throws error when provider is missing", async () => {
-        const output = await api.getProviderTokens("unsuppported", { headers: new Headers() })
+        const output = await api.getAccessToken("unsuppported", { headers: new Headers() })
         expect(output).toEqual({
             success: false,
-            tokens: null,
+            accessToken: null,
             error: {
                 code: "UNSUPPORTED_OAUTH_CONFIGURATION",
                 message: "The targeted OAuth provider has not been configured in the initialization parameters.",
@@ -31,10 +25,10 @@ describe("getProviderTokens API", () => {
     })
 
     test("throws error when session token is missing", async () => {
-        const output = await api.getProviderTokens("oauth-provider", { headers: new Headers() })
+        const output = await api.getAccessToken("oauth-provider", { headers: new Headers() })
         expect(output).toEqual({
             success: false,
-            tokens: null,
+            accessToken: null,
             error: {
                 code: "SESSION_NOT_FOUND",
                 message: "The session token is not found. There is no active session.",
@@ -47,14 +41,14 @@ describe("getProviderTokens API", () => {
     test("throws error when CSRF token is missing", async () => {
         const sessionToken = await jose.encodeJWT(sessionPayload)
 
-        const output = await api.getProviderTokens("oauth-provider", {
+        const output = await api.getAccessToken("oauth-provider", {
             headers: {
                 Cookie: `aura-auth.session_token=${sessionToken}`,
             },
         })
         expect(output).toEqual({
             success: false,
-            tokens: null,
+            accessToken: null,
             error: {
                 code: "CSRF_TOKEN_MISSING",
                 message: "The CSRF token is missing. Please refresh and try again.",
@@ -68,7 +62,7 @@ describe("getProviderTokens API", () => {
         vi.stubEnv("BASE_URL", "https://example.com")
         const sessionToken = await jose.encodeJWT(sessionPayload)
 
-        const output = await api.getProviderTokens("oauth-provider", {
+        const output = await api.getAccessToken("oauth-provider", {
             headers: new Headers({
                 Cookie: `aura-auth.csrf_token=invalid-token; aura-auth.session_token=${sessionToken}`,
                 "X-CSRF-Token": "invalid-token",
@@ -76,7 +70,7 @@ describe("getProviderTokens API", () => {
         })
         expect(output).toEqual({
             success: false,
-            tokens: null,
+            accessToken: null,
             error: {
                 code: "CSRF_TOKEN_MISMATCH",
                 message: "CSRF token verification failed. Please refresh and try again.",
@@ -91,7 +85,7 @@ describe("getProviderTokens API", () => {
         const csrfToken = await createCSRF(jose)
         const sessionToken = await jose.encodeJWT(sessionPayload)
 
-        const output = await api.getProviderTokens("oauth-provider", {
+        const output = await api.getAccessToken("oauth-provider", {
             headers: {
                 "X-CSRF-Token": csrfToken,
                 Cookie: `aura-auth.csrf_token=${csrfToken}; aura-auth.session_token=${sessionToken}`,
@@ -99,7 +93,7 @@ describe("getProviderTokens API", () => {
         })
         expect(output).toEqual({
             success: false,
-            tokens: null,
+            accessToken: null,
             error: {
                 code: "COOKIE_INVALID_VALUE",
                 message: "Expected configuration cookie not found or contains an empty value.",
@@ -116,7 +110,7 @@ describe("getProviderTokens API", () => {
 
         const encodedTokens = await jose.encodeJWT(oauthTokens as unknown as Record<string, unknown>)
 
-        const output = await api.getProviderTokens("oauth-provider", {
+        const output = await api.getAccessToken("oauth-provider", {
             headers: {
                 "X-CSRF-Token": csrfToken,
                 Cookie: `aura-auth.csrf_token=${csrfToken}; aura-auth.session_token=${sessionToken}; aura-auth.access_token.oauth-provider=${encodedTokens}`,
@@ -124,7 +118,7 @@ describe("getProviderTokens API", () => {
         })
         expect(output).toEqual({
             success: true,
-            tokens: oauthTokens,
+            accessToken: oauthTokens.accessToken,
             headers: expect.any(Headers),
             toResponse: expect.any(Function),
         })
@@ -143,7 +137,7 @@ describe("getProviderTokens API", () => {
         const { refreshToken: _, ...spread } = oauthCustomService
         const { api } = createAuth({ oauth: [spread] })
 
-        const output = await api.getProviderTokens("oauth-provider", {
+        const output = await api.getAccessToken("oauth-provider", {
             headers: {
                 "X-CSRF-Token": csrfToken,
                 Cookie: `aura-auth.csrf_token=${csrfToken}; aura-auth.session_token=${sessionToken}; aura-auth.access_token.oauth-provider=${encodedTokens}`,
@@ -151,7 +145,7 @@ describe("getProviderTokens API", () => {
         })
         expect(output).toEqual({
             success: false,
-            tokens: null,
+            accessToken: null,
             error: {
                 code: "OAUTH_INVALID_REFRESH_TOKEN_CONFIG",
                 message:
@@ -181,7 +175,7 @@ describe("getProviderTokens API", () => {
             expiresAt: Math.floor(Date.now() / 1000) - 3600,
         } as unknown as Record<string, unknown>)
 
-        const output = await api.getProviderTokens("oauth-provider", {
+        const output = await api.getAccessToken("oauth-provider", {
             headers: {
                 "X-CSRF-Token": csrfToken,
                 Cookie: `aura-auth.csrf_token=${csrfToken}; aura-auth.session_token=${sessionToken}; aura-auth.access_token.oauth-provider=${encodedTokens}`,
@@ -189,11 +183,7 @@ describe("getProviderTokens API", () => {
         })
         expect(output).toEqual({
             success: true,
-            tokens: {
-                ...oauthTokens,
-                expiresAt: expect.any(Number),
-                issuedAt: expect.any(Number),
-            },
+            accessToken: oauthTokens.accessToken,
             headers: expect.any(Headers),
             toResponse: expect.any(Function),
         })
@@ -243,7 +233,7 @@ describe("getProviderTokens API", () => {
             oauth: [provider],
         })
 
-        const output = await api.getProviderTokens("oauth-provider", {
+        const output = await api.getAccessToken("oauth-provider", {
             headers: {
                 "X-CSRF-Token": csrfToken,
                 Cookie: `aura-auth.csrf_token=${csrfToken}; aura-auth.session_token=${sessionToken}; aura-auth.access_token.oauth-provider=${encodedTokens}`,
@@ -251,11 +241,7 @@ describe("getProviderTokens API", () => {
         })
         expect(output).toEqual({
             success: true,
-            tokens: {
-                ...oauthTokens,
-                expiresAt: expect.any(Number),
-                issuedAt: expect.any(Number),
-            },
+            accessToken: oauthTokens.accessToken,
             headers: expect.any(Headers),
             toResponse: expect.any(Function),
         })
@@ -292,7 +278,7 @@ describe("getProviderTokens API", () => {
             expiresAt: Math.floor(Date.now() / 1000) - 3600,
         } as unknown as Record<string, unknown>)
 
-        const output = await api.getProviderTokens("oauth-provider", {
+        const output = await api.getAccessToken("oauth-provider", {
             headers: {
                 "X-CSRF-Token": csrfToken,
                 Cookie: `aura-auth.csrf_token=${csrfToken}; aura-auth.session_token=${sessionToken}; aura-auth.access_token.oauth-provider=${encodedTokens}`,
@@ -301,7 +287,7 @@ describe("getProviderTokens API", () => {
 
         expect(output).toEqual({
             success: false,
-            tokens: null,
+            accessToken: null,
             error: {
                 code: "OAUTH_INVALID_REFRESH_TOKEN_RESPONSE",
                 message: "Your secure session renewal failed. Please sign in again to continue.",
@@ -324,7 +310,7 @@ describe("getProviderTokens API", () => {
             expiresAt: Math.floor(Date.now() / 1000) - 3600,
         } as unknown as Record<string, unknown>)
 
-        const output = await api.getProviderTokens("oauth-provider", {
+        const output = await api.getAccessToken("oauth-provider", {
             headers: {
                 "X-CSRF-Token": csrfToken,
                 Cookie: `aura-auth.csrf_token=${csrfToken}; aura-auth.session_token=${sessionToken}; aura-auth.access_token.oauth-provider=${encodedTokens}`,
@@ -333,7 +319,7 @@ describe("getProviderTokens API", () => {
 
         expect(output).toEqual({
             success: false,
-            tokens: null,
+            accessToken: null,
             error: {
                 code: "PROVIDER_TOKENS_ERROR",
                 message: "Failed to get provider tokens",
@@ -357,7 +343,7 @@ describe("getProviderTokens API", () => {
             expiresAt: currentTime + 600,
         } as unknown as Record<string, unknown>)
 
-        const output = await api.getProviderTokens("oauth-provider", {
+        const output = await api.getAccessToken("oauth-provider", {
             headers: {
                 "X-CSRF-Token": csrfToken,
                 Cookie: `aura-auth.csrf_token=${csrfToken}; aura-auth.session_token=${sessionToken}; aura-auth.access_token.oauth-provider=${encodedTokens}`,
@@ -366,10 +352,7 @@ describe("getProviderTokens API", () => {
 
         expect(output).toEqual({
             success: true,
-            tokens: {
-                ...oauthTokens,
-                expiresAt: currentTime + 600,
-            },
+            accessToken: oauthTokens.accessToken,
             headers: expect.any(Headers),
             toResponse: expect.any(Function),
         })
@@ -398,7 +381,7 @@ describe("getProviderTokens API", () => {
             expiresAt: currentTime + 120,
         } as unknown as Record<string, unknown>)
 
-        const output = await api.getProviderTokens("oauth-provider", {
+        const output = await api.getAccessToken("oauth-provider", {
             headers: {
                 "X-CSRF-Token": csrfToken,
                 Cookie: `aura-auth.csrf_token=${csrfToken}; aura-auth.session_token=${sessionToken}; aura-auth.access_token.oauth-provider=${encodedTokens}`,
@@ -407,44 +390,10 @@ describe("getProviderTokens API", () => {
 
         expect(output).toEqual({
             success: true,
-            tokens: {
-                ...oauthTokens,
-                accessToken: "brand-new-refreshed-token",
-                expiresAt: expect.any(Number),
-                issuedAt: expect.any(Number),
-            },
+            accessToken: "brand-new-refreshed-token",
             headers: expect.any(Headers),
             toResponse: expect.any(Function),
         })
         expect(mockFetch).toHaveBeenCalledTimes(1)
-    })
-
-    test("missing expires_in", async () => {
-        vi.stubEnv("BASE_URL", "http://localhost:3000")
-        const csrfToken = await createCSRF(jose)
-        const sessionToken = await jose.encodeJWT(sessionPayload)
-
-        const encodedTokens = await jose.encodeJWT({
-            accessToken: "access-token",
-            scopes: ["read:user,read:email"],
-            tokenType: "bearer",
-        } as unknown as Record<string, unknown>)
-
-        const output = await api.getProviderTokens("oauth-provider", {
-            headers: {
-                "X-CSRF-Token": csrfToken,
-                Cookie: `aura-auth.csrf_token=${csrfToken}; aura-auth.session_token=${sessionToken}; aura-auth.access_token.oauth-provider=${encodedTokens}`,
-            },
-        })
-        expect(output).toEqual({
-            success: true,
-            tokens: {
-                accessToken: "access-token",
-                scopes: ["read:user,read:email"],
-                tokenType: "Bearer",
-            },
-            headers: expect.any(Headers),
-            toResponse: expect.any(Function),
-        })
     })
 })
