+++ packages/core/test/api/stateless/getProviderTokens.test.ts
@@ -1,4 +1,4 @@
-import { describe, test, expect, vi, beforeEach, afterEach } from "vitest"
+import { describe, test, expect, vi, beforeEach } from "vitest"
 import { api, jose, oauthCustomService, oauthTokens, sessionPayload } from "@test/presets.ts"
 import { createCSRF } from "@/shared/crypto.ts"
 import { createAuth } from "@/createAuth.ts"
@@ -9,16 +9,12 @@ beforeEach(() => {
     vi.stubEnv("BASE_URL", undefined)
 })
 
-afterEach(() => {
-    vi.unstubAllEnvs()
-})
-
-describe("getAccessToken API", () => {
+describe("getProviderTokens API", () => {
     test("throws error when provider is missing", async () => {
-        const output = await api.getAccessToken("unsuppported", { headers: new Headers() })
+        const output = await api.getProviderTokens("unsuppported", { headers: new Headers() })
         expect(output).toEqual({
             success: false,
-            accessToken: null,
+            tokens: null,
             error: {
                 code: "UNSUPPORTED_OAUTH_CONFIGURATION",
                 message: "The targeted OAuth provider has not been configured in the initialization parameters.",
@@ -29,10 +25,10 @@ describe("getAccessToken API", () => {
     })
 
     test("throws error when session token is missing", async () => {
-        const output = await api.getAccessToken("oauth-provider", { headers: new Headers() })
+        const output = await api.getProviderTokens("oauth-provider", { headers: new Headers() })
         expect(output).toEqual({
             success: false,
-            accessToken: null,
+            tokens: null,
             error: {
                 code: "SESSION_NOT_FOUND",
                 message: "The session token is not found. There is no active session.",
@@ -45,14 +41,14 @@ describe("getAccessToken API", () => {
     test("throws error when CSRF token is missing", async () => {
         const sessionToken = await jose.encodeJWT(sessionPayload)
 
-        const output = await api.getAccessToken("oauth-provider", {
+        const output = await api.getProviderTokens("oauth-provider", {
             headers: {
                 Cookie: `aura-auth.session_token=${sessionToken}`,
             },
         })
         expect(output).toEqual({
             success: false,
-            accessToken: null,
+            tokens: null,
             error: {
                 code: "CSRF_TOKEN_MISSING",
                 message: "The CSRF token is missing. Please refresh and try again.",
@@ -66,7 +62,7 @@ describe("getAccessToken API", () => {
         vi.stubEnv("BASE_URL", "https://example.com")
         const sessionToken = await jose.encodeJWT(sessionPayload)
 
-        const output = await api.getAccessToken("oauth-provider", {
+        const output = await api.getProviderTokens("oauth-provider", {
             headers: new Headers({
                 Cookie: `aura-auth.csrf_token=invalid-token; aura-auth.session_token=${sessionToken}`,
                 "X-CSRF-Token": "invalid-token",
@@ -74,7 +70,7 @@ describe("getAccessToken API", () => {
         })
         expect(output).toEqual({
             success: false,
-            accessToken: null,
+            tokens: null,
             error: {
                 code: "CSRF_TOKEN_MISMATCH",
                 message: "CSRF token verification failed. Please refresh and try again.",
@@ -89,7 +85,7 @@ describe("getAccessToken API", () => {
         const csrfToken = await createCSRF(jose)
         const sessionToken = await jose.encodeJWT(sessionPayload)
 
-        const output = await api.getAccessToken("oauth-provider", {
+        const output = await api.getProviderTokens("oauth-provider", {
             headers: {
                 "X-CSRF-Token": csrfToken,
                 Cookie: `aura-auth.csrf_token=${csrfToken}; aura-auth.session_token=${sessionToken}`,
@@ -97,7 +93,7 @@ describe("getAccessToken API", () => {
         })
         expect(output).toEqual({
             success: false,
-            accessToken: null,
+            tokens: null,
             error: {
                 code: "COOKIE_INVALID_VALUE",
                 message: "Expected configuration cookie not found or contains an empty value.",
@@ -114,7 +110,7 @@ describe("getAccessToken API", () => {
 
         const encodedTokens = await jose.encodeJWT(oauthTokens as unknown as Record<string, unknown>)
 
-        const output = await api.getAccessToken("oauth-provider", {
+        const output = await api.getProviderTokens("oauth-provider", {
             headers: {
                 "X-CSRF-Token": csrfToken,
                 Cookie: `aura-auth.csrf_token=${csrfToken}; aura-auth.session_token=${sessionToken}; aura-auth.access_token.oauth-provider=${encodedTokens}`,
@@ -122,7 +118,7 @@ describe("getAccessToken API", () => {
         })
         expect(output).toEqual({
             success: true,
-            accessToken: oauthTokens.accessToken,
+            tokens: oauthTokens,
             headers: expect.any(Headers),
             toResponse: expect.any(Function),
         })
@@ -141,7 +137,7 @@ describe("getAccessToken API", () => {
         const { refreshToken: _, ...spread } = oauthCustomService
         const { api } = createAuth({ oauth: [spread] })
 
-        const output = await api.getAccessToken("oauth-provider", {
+        const output = await api.getProviderTokens("oauth-provider", {
             headers: {
                 "X-CSRF-Token": csrfToken,
                 Cookie: `aura-auth.csrf_token=${csrfToken}; aura-auth.session_token=${sessionToken}; aura-auth.access_token.oauth-provider=${encodedTokens}`,
@@ -149,7 +145,7 @@ describe("getAccessToken API", () => {
         })
         expect(output).toEqual({
             success: false,
-            accessToken: null,
+            tokens: null,
             error: {
                 code: "OAUTH_INVALID_REFRESH_TOKEN_CONFIG",
                 message:
@@ -179,7 +175,7 @@ describe("getAccessToken API", () => {
             expiresAt: Math.floor(Date.now() / 1000) - 3600,
         } as unknown as Record<string, unknown>)
 
-        const output = await api.getAccessToken("oauth-provider", {
+        const output = await api.getProviderTokens("oauth-provider", {
             headers: {
                 "X-CSRF-Token": csrfToken,
                 Cookie: `aura-auth.csrf_token=${csrfToken}; aura-auth.session_token=${sessionToken}; aura-auth.access_token.oauth-provider=${encodedTokens}`,
@@ -187,7 +183,11 @@ describe("getAccessToken API", () => {
         })
         expect(output).toEqual({
             success: true,
-            accessToken: oauthTokens.accessToken,
+            tokens: {
+                ...oauthTokens,
+                expiresAt: expect.any(Number),
+                issuedAt: expect.any(Number),
+            },
             headers: expect.any(Headers),
             toResponse: expect.any(Function),
         })
@@ -237,7 +237,7 @@ describe("getAccessToken API", () => {
             oauth: [provider],
         })
 
-        const output = await api.getAccessToken("oauth-provider", {
+        const output = await api.getProviderTokens("oauth-provider", {
             headers: {
                 "X-CSRF-Token": csrfToken,
                 Cookie: `aura-auth.csrf_token=${csrfToken}; aura-auth.session_token=${sessionToken}; aura-auth.access_token.oauth-provider=${encodedTokens}`,
@@ -245,7 +245,11 @@ describe("getAccessToken API", () => {
         })
         expect(output).toEqual({
             success: true,
-            accessToken: oauthTokens.accessToken,
+            tokens: {
+                ...oauthTokens,
+                expiresAt: expect.any(Number),
+                issuedAt: expect.any(Number),
+            },
             headers: expect.any(Headers),
             toResponse: expect.any(Function),
         })
@@ -282,7 +286,7 @@ describe("getAccessToken API", () => {
             expiresAt: Math.floor(Date.now() / 1000) - 3600,
         } as unknown as Record<string, unknown>)
 
-        const output = await api.getAccessToken("oauth-provider", {
+        const output = await api.getProviderTokens("oauth-provider", {
             headers: {
                 "X-CSRF-Token": csrfToken,
                 Cookie: `aura-auth.csrf_token=${csrfToken}; aura-auth.session_token=${sessionToken}; aura-auth.access_token.oauth-provider=${encodedTokens}`,
@@ -291,7 +295,7 @@ describe("getAccessToken API", () => {
 
         expect(output).toEqual({
             success: false,
-            accessToken: null,
+            tokens: null,
             error: {
                 code: "OAUTH_INVALID_REFRESH_TOKEN_RESPONSE",
                 message: "Your secure session renewal failed. Please sign in again to continue.",
@@ -314,7 +318,7 @@ describe("getAccessToken API", () => {
             expiresAt: Math.floor(Date.now() / 1000) - 3600,
         } as unknown as Record<string, unknown>)
 
-        const output = await api.getAccessToken("oauth-provider", {
+        const output = await api.getProviderTokens("oauth-provider", {
             headers: {
                 "X-CSRF-Token": csrfToken,
                 Cookie: `aura-auth.csrf_token=${csrfToken}; aura-auth.session_token=${sessionToken}; aura-auth.access_token.oauth-provider=${encodedTokens}`,
@@ -323,7 +327,7 @@ describe("getAccessToken API", () => {
 
         expect(output).toEqual({
             success: false,
-            accessToken: null,
+            tokens: null,
             error: {
                 code: "PROVIDER_TOKENS_ERROR",
                 message: "Failed to get provider tokens",
@@ -347,7 +351,7 @@ describe("getAccessToken API", () => {
             expiresAt: currentTime + 600,
         } as unknown as Record<string, unknown>)
 
-        const output = await api.getAccessToken("oauth-provider", {
+        const output = await api.getProviderTokens("oauth-provider", {
             headers: {
                 "X-CSRF-Token": csrfToken,
                 Cookie: `aura-auth.csrf_token=${csrfToken}; aura-auth.session_token=${sessionToken}; aura-auth.access_token.oauth-provider=${encodedTokens}`,
@@ -356,7 +360,10 @@ describe("getAccessToken API", () => {
 
         expect(output).toEqual({
             success: true,
-            accessToken: oauthTokens.accessToken,
+            tokens: {
+                ...oauthTokens,
+                expiresAt: currentTime + 600,
+            },
             headers: expect.any(Headers),
             toResponse: expect.any(Function),
         })
@@ -385,7 +392,7 @@ describe("getAccessToken API", () => {
             expiresAt: currentTime + 120,
         } as unknown as Record<string, unknown>)
 
-        const output = await api.getAccessToken("oauth-provider", {
+        const output = await api.getProviderTokens("oauth-provider", {
             headers: {
                 "X-CSRF-Token": csrfToken,
                 Cookie: `aura-auth.csrf_token=${csrfToken}; aura-auth.session_token=${sessionToken}; aura-auth.access_token.oauth-provider=${encodedTokens}`,
@@ -394,10 +401,44 @@ describe("getAccessToken API", () => {
 
         expect(output).toEqual({
             success: true,
-            accessToken: "brand-new-refreshed-token",
+            tokens: {
+                ...oauthTokens,
+                accessToken: "brand-new-refreshed-token",
+                expiresAt: expect.any(Number),
+                issuedAt: expect.any(Number),
+            },
             headers: expect.any(Headers),
             toResponse: expect.any(Function),
         })
         expect(mockFetch).toHaveBeenCalledTimes(1)
+    })
+
+    test("missing expires_in", async () => {
+        vi.stubEnv("BASE_URL", "http://localhost:3000")
+        const csrfToken = await createCSRF(jose)
+        const sessionToken = await jose.encodeJWT(sessionPayload)
+
+        const encodedTokens = await jose.encodeJWT({
+            accessToken: "access-token",
+            scopes: ["read:user,read:email"],
+            tokenType: "bearer",
+        } as unknown as Record<string, unknown>)
+
+        const output = await api.getProviderTokens("oauth-provider", {
+            headers: {
+                "X-CSRF-Token": csrfToken,
+                Cookie: `aura-auth.csrf_token=${csrfToken}; aura-auth.session_token=${sessionToken}; aura-auth.access_token.oauth-provider=${encodedTokens}`,
+            },
+        })
+        expect(output).toEqual({
+            success: true,
+            tokens: {
+                accessToken: "access-token",
+                scopes: ["read:user,read:email"],
+                tokenType: "Bearer",
+            },
+            headers: expect.any(Headers),
+            toResponse: expect.any(Function),
+        })
     })
 })
