+++ packages/core/test/api/stateless/signIn.test.ts
@@ -1,37 +1,10 @@
-import { describe, test, expect, vi, beforeEach, afterEach } from "vitest"
+import { describe, test, expect, vi, beforeEach } from "vitest"
 import { getSetCookie } from "@/cookie.ts"
 import { createAuth } from "@/createAuth.ts"
 import { api, oauthCustomService, openIDMetadata } from "@test/presets.ts"
 
 beforeEach(() => {
     vi.stubEnv("BASE_URL", undefined)
-})
-
-afterEach(() => {
-    vi.unstubAllEnvs()
-})
-
-vi.mock("@aura-stack/rate-limiter", async () => {
-    const actual = await vi.importActual<typeof import("@aura-stack/rate-limiter")>("@aura-stack/rate-limiter")
-    return {
-        ...actual,
-        createRateLimiter: (...args: Parameters<typeof actual.createRateLimiter>) => {
-            const limiters = actual.createRateLimiter(...args)
-
-            for (const limiter of Object.values(limiters)) {
-                limiter.check = vi.fn().mockResolvedValue({
-                    ok: true,
-                    limit: Number.MAX_SAFE_INTEGER,
-                    remaining: Number.MAX_SAFE_INTEGER,
-                    resetAt: Date.now() + 60000,
-                    retryAfter: 0,
-                    toResponse: () => new Response(),
-                })
-            }
-
-            return limiters
-        },
-    }
 })
 
 describe("signIn API", () => {
