+++ packages/core/test/instance.test.ts
@@ -1,12 +1,8 @@
-import { describe, test, expect, vi, beforeEach, afterEach } from "vitest"
+import { describe, test, expect, vi, beforeEach } from "vitest"
 import { createAuth } from "@/createAuth.ts"
 
 beforeEach(() => {
     vi.stubEnv("BASE_URL", undefined)
-})
-
-afterEach(() => {
-    vi.unstubAllEnvs()
 })
 
 describe("createAuth", () => {
