+++ packages/core/test/setup/vitest.setup.ts
@@ -1,0 +1,31 @@
+import { vi, afterEach } from "vitest"
+
+afterEach(() => {
+    vi.clearAllMocks()
+    vi.unstubAllEnvs()
+    vi.restoreAllMocks()
+    vi.unstubAllGlobals()
+})
+
+vi.mock("@aura-stack/rate-limiter", async () => {
+    const actual = await vi.importActual<typeof import("@aura-stack/rate-limiter")>("@aura-stack/rate-limiter")
+    return {
+        ...actual,
+        createRateLimiter: (...args: Parameters<typeof actual.createRateLimiter>) => {
+            const limiters = actual.createRateLimiter(...args)
+
+            for (const limiter of Object.values(limiters)) {
+                limiter.check = vi.fn().mockResolvedValue({
+                    ok: true,
+                    limit: Number.MAX_SAFE_INTEGER,
+                    remaining: Number.MAX_SAFE_INTEGER,
+                    resetAt: Date.now() + 60000,
+                    retryAfter: 0,
+                    toResponse: () => new Response(),
+                })
+            }
+
+            return limiters
+        },
+    }
+})
